+++ spec/array_spec.rb
@@ -3,6 +3,9 @@
 RSpec.describe Shaped::Array do
   subject(:shaped_array) { Shaped::Array(shape_description) }
 
+  let(:shape_description) { [Numeric] }
+  let(:test_array) { [1, 2.2, 1_024] }
+
   describe '#initialize' do
     context 'when the shape description is an array of one class' do
       let(:shape_description) { [String] }
@@ -39,7 +42,7 @@ RSpec.describe Shaped::Array do
     subject(:matched_by?) { shaped_array.matched_by?(test_array) }
 
     context 'when the shape description is `[Numeric]`' do
-      let(:shape_description) { [Numeric] }
+      before { expect(shape_description).to eq([Numeric]) }
 
       context 'when the test array is an empty array' do
         let(:test_array) { [] }
@@ -70,7 +73,7 @@ RSpec.describe Shaped::Array do
       end
 
       context 'when the test array is an array of solely Numerics' do
-        let(:test_array) { [1, 2.2, 1_024] }
+        before { expect(test_array).to eq([1, 2.2, 1_024]) }
 
         it 'returns true' do
           expect(matched_by?).to eq(true)
@@ -83,6 +86,31 @@ RSpec.describe Shaped::Array do
         it 'returns false' do
           expect(matched_by?).to eq(false)
         end
+      end
+    end
+  end
+
+  describe '#match_failure_reason' do
+    subject(:match_failure_reason) { shaped_array.match_failure_reason(test_array) }
+
+    context 'when the test array matches the shape description' do
+      before { expect(shaped_array).to be_matched_by(test_array) }
+
+      it 'returns nil' do
+        expect(match_failure_reason).to eq(nil)
+      end
+    end
+
+    context 'when the test array does not match the shape description' do
+      before { expect(shaped_array).not_to be_matched_by(test_array) }
+
+      let(:test_array) { [1, 2.0, 'not a number!', Rational(3, 4)] }
+
+      it 'returns a Shaped::MatchFailureReason instance with the correct info' do
+        expect(match_failure_reason).to be_a(Shaped::MatchFailureReason)
+        expect(match_failure_reason.to_s).to eq(
+          'Object at `2` is expected to be a Numeric but was a String',
+        )
       end
     end
   end